Remove like and rc4 in ShardingSphere encrypt and move to ShardingSphere plugin repository
