+++ features/encrypt/distsql/parser/src/main/antlr4/imports/encrypt/BaseRule.g4
@@ -34,23 +34,14 @@ algorithmTypeName
 buildInEncryptAlgorithmType
     : standardEncryptAlgorithm
     | assistedEncryptAlgorithm
-    | likeEncryptAlgorithm
     ;
 
 standardEncryptAlgorithm
-    : MD5
-    | AES
-    | RC4
-    | SM3
-    | SM4
+    : AES
     ;
 
 assistedEncryptAlgorithm
     : MD5
-    ;
-
-likeEncryptAlgorithm
-    : CHAR_DIGEST_LIKE
     ;
 
 propertiesDefinition
