+++ features/encrypt/distsql/parser/src/main/antlr4/imports/encrypt/Keyword.g4
@@ -143,22 +143,6 @@ AES
     : A E S
     ;
 
-RC4
-    : R C [4]
-    ;
-
-SM3
-    : S M [3]
-    ;
-
-SM4
-    : S M [4]
-    ;
-
-CHAR_DIGEST_LIKE
-    : C H A R UL_ D I G E S T UL_ L I K E
-    ;
-
 NOT
     : N O T
     ;
