+++ test/e2e/fixture/src/test/resources/algorithm/like/common_chinese_character.dict
