+++ test/e2e/sql/src/test/resources/cases/rql/dataset/encrypt/show_encrypt_rule.xml
@@ -29,8 +29,8 @@
         <column name="like_query_type" />
         <column name="like_query_props" />
     </metadata>
-    <row values="t_user| user_name| user_name_cipher| | user_name_like| AES| aes-key-value=123456abc| | | CHAR_DIGEST_LIKE| mask=4093" />
+    <row values="t_user| user_name| user_name_cipher| | user_name_like| AES| aes-key-value=123456abc| | | IT.ENCRYPT.LIKE.FIXTURE| mask=4093" />
     <row values="t_user| password| password_cipher| | | AES| aes-key-value=123456abc| | | | " />
     <row values="t_user| email| email_cipher| | | AES| aes-key-value=123456abc| | | | " />
-    <row values="t_user| telephone| user_telephone_cipher| | user_telephone_like| AES| aes-key-value=123456abc| | | CHAR_DIGEST_LIKE| mask=4093" />
+    <row values="t_user| telephone| user_telephone_cipher| | user_telephone_like| AES| aes-key-value=123456abc| | | IT.ENCRYPT.LIKE.FIXTURE| mask=4093" />
 </dataset>
