+++ test/e2e/sql/src/test/resources/cases/rql/dataset/encrypt/show_encrypt_rules.xml
@@ -29,12 +29,12 @@
         <column name="like_query_type" />
         <column name="like_query_props" />
     </metadata>
-    <row values="t_user| user_name| user_name_cipher| | user_name_like| AES| aes-key-value=123456abc| | | CHAR_DIGEST_LIKE| mask=4093" />
+    <row values="t_user| user_name| user_name_cipher| | user_name_like| AES| aes-key-value=123456abc| | | IT.ENCRYPT.LIKE.FIXTURE| mask=4093" />
     <row values="t_user| password| password_cipher| | | AES| aes-key-value=123456abc| | | | " />
     <row values="t_user| email| email_cipher| | | AES| aes-key-value=123456abc| | | | " />
-    <row values="t_user| telephone| user_telephone_cipher| | user_telephone_like| AES| aes-key-value=123456abc| | | CHAR_DIGEST_LIKE| mask=4093" />
+    <row values="t_user| telephone| user_telephone_cipher| | user_telephone_like| AES| aes-key-value=123456abc| | | IT.ENCRYPT.LIKE.FIXTURE| mask=4093" />
     <row values="t_user_details| number| number_cipher| | | AES| aes-key-value=123456abc| | | | " />
     <row values="t_user_details| number_new| number_new_cipher| | | AES| aes-key-value=123456abc| | | | " />
-    <row values="t_merchant| business_code| business_code_cipher| | business_code_like| AES| aes-key-value=123456abc| | | CHAR_DIGEST_LIKE | mask=4093 " />
-    <row values="t_merchant| telephone| merchant_telephone_cipher| | merchant_telephone_like| AES| aes-key-value=123456abc| | | CHAR_DIGEST_LIKE | mask=4093 " />
+    <row values="t_merchant| business_code| business_code_cipher| | business_code_like| AES| aes-key-value=123456abc| | | IT.ENCRYPT.LIKE.FIXTURE | mask=4093 " />
+    <row values="t_merchant| telephone| merchant_telephone_cipher| | merchant_telephone_like| AES| aes-key-value=123456abc| | | IT.ENCRYPT.LIKE.FIXTURE | mask=4093 " />
 </dataset>
